Refactoring: Extracted ArmyListLoader
